ADD: template page, template select in plu update form, proxy production site in common select

+++ Src/Apps/Web/Ws.DeviceControl.Models/Dto/References/Template/TemplateMapper.cs
@@ -10,9 +10,9 @@ public static class TemplateMapper
         return new()
         {
             Name = item.Name,
-            Width = item.Width,
-            Height = item.Height,
-            Rotate = item.Rotate,
+            Width = (short) item.Width,
+            Height = (short) item.Height,
+            Rotate = (short) item.Rotate,
         };
     }
 }
